# perf: apply memory optimizations to `TokenReader`